Fix head solver issues

+++ Solver/Solver/HeadSolver.swift
@@ -1,12 +1,12 @@
 struct HeadSolver: Solver {
     func solve(problem: Problem) -> Solution {
-        var sum = 0 as UInt
-        var indices = [UInt32]()
+        var sum = 0 as UInt64
+        var indices = [Int]()
         
         for (idx, pizza) in problem.pizzas.enumerated() {
             sum += pizza.numberOfSlices
             if sum < problem.maximumNumberOfSlices {
-                indices.append(UInt32(idx))
+                indices.append(idx)
             } else {
                 break
             }
